rename main file to internal_coords, toplevel commands file to ic_rebuild
